backend/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "fastapi[standard]>=0.115.7",
    "jq>=1.8.0",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.2",
    "langchain-pinecone>=0.2.3",
    "langgraph>=0.3.5",
    "pandas>=2.2.3",
    "pinecone-notebooks>=0.1.1",
    "plotly>=6.0.0",
    "pre-commit>=4.1.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-socketio>=5.12.1",
    "ruff>=0.9.6",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]
